catfish_simple/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
